requirements.txt
PyQt6>=6.4.0
geopy>=2.3.0
ephem>=4.1.4
pyswisseph>=2.10.3.1
pytz>=2023.3
timezonefinder>=6.5.5
kerykeion>=0.1.0
pandas>=1.3.0